requirements.txt
robotframework==5.0.1
robotframework-pabot==2.18.0
robotframework-seleniumlibrary==5.1.3
selenium==4.9.0
allure-robotframework==2.13.2
